fix: 매니저 마이페이지 내 정보 조회 수정
